+++ package.json
@@ -31,7 +31,7 @@
     "@polkadot/react-identicon": "^3.2.2",
     "@polkadot/rpc-provider": "^10.5.1",
     "@polkadot/util": "^12.1.1",
-    "@polkadotcloud/core-ui": "0.2.33",
+    "@polkadotcloud/core-ui": "0.2.34",
     "@polkadotcloud/react-odometer": "0.1.16",
     "@polkadotcloud/utils": "0.2.19",
     "@substrate/connect": "^0.7.24",
